Created the Channel component and refactored App.js to render the Channel component
